test version

+++ controllers/majorItem.controller.js
@@ -1,3 +1,4 @@
+import Center from "../models/center.model.js";
 import MajorItem from "../models/majorItem.model.js";
 import majorItemValid from "../validations/majorItem.valid.js";
 import { Op } from "sequelize";
@@ -9,6 +10,15 @@ export async function create(req, res) {
          return res.status(422).json({ message: error.details[0].message });
       }
       let { majorItems } = req.body;
+      
+      let center = await Center.findByPk(majorItems[0].centerId);
+      if (!center) {
+         return res.status(404).json({ message: "Not found centerId." });
+      }
+
+      if (center.seoId != req.user.id && req.user.role != "admin") {
+         return res.status(401).json({ message: "Not allowed." });
+      }
 
       let majorItem = await MajorItem.bulkCreate(majorItems.map((x) => x));
 
@@ -21,6 +31,15 @@ export async function create(req, res) {
 export async function remove(req, res) {
    try {
       let { centerId, majorItems } = req.body;
+
+      let center = await Center.findByPk(centerId);
+      if (!center) {
+         return res.status(404).json({ message: "Not found centerId." });
+      }
+
+      if (center.seoId != req.user.id && req.user.role != "admin") {
+         return res.status(401).json({ message: "Not allowed." });
+      }
 
       for (let majorId of majorItems) {
          let del = await MajorItem.destroy({
